+++ stdnet/backends/base.py
@@ -93,17 +93,17 @@ class BackendStructure(AsyncObject):
     def backend_structure(self):
         return self
     
-    def delete(self):
-        raise NotImplementedError()
-    
-    def flush(self):
-        raise NotImplementedError()
-    
-    def size(self):
-        raise NotImplementedError()
-    
     def clone(self):
         return self.__class__(self.instance,self.client)
+    
+    def delete(self):   # pragma: no cover
+        raise NotImplementedError()
+    
+    def flush(self):    # pragma: no cover
+        raise NotImplementedError()
+    
+    def size(self):     # pragma: no cover
+        raise NotImplementedError()
     
     
 class BackendQuery(ServerOperation):