+++ tests/regression/serialize/base.py
@@ -1,6 +1,4 @@
-__test__ = False
-
-from stdnet import orm
+from stdnet import orm, test
 
 
 class SerializerMixin(object):
@@ -8,7 +6,8 @@ class SerializerMixin(object):
     
     def get(self, **options):
         s = orm.get_serializer(self.serializer)
-        self.assertEqual(s.options,options)
+        if not s.default_options:
+            self.assertEqual(s.options, options)
         self.assertFalse(s.data)
         self.assertTrue(s)
         return s
@@ -34,4 +33,25 @@ class SerializerMixin(object):
         data = s.write().getvalue()
         self.model.objects.flush()
         s.load(data)
+
+
+class DummySerializer(orm.Serializer):
+    pass
+
+
+class TestMeta(test.TestCase):
+    
+    def testBadSerializer(self):
+        self.assertRaises(ValueError, orm.get_serializer, 'djsbvjchvsdjcvsdj')
+        
+    def testRegisterUnregister(self):
+        orm.register_serializer('dummy',DummySerializer())
+        s = orm.get_serializer('dummy')
+        self.assertTrue('dummy' in orm.all_serializers())
+        self.assertTrue(isinstance(s,DummySerializer))
+        self.assertRaises(NotImplementedError, s.serialize, None)
+        self.assertRaises(NotImplementedError, s.write)
+        self.assertRaises(NotImplementedError, s.load, None)
+        self.assertTrue(orm.unregister_serializer('dummy'))
+        self.assertRaises(ValueError, orm.get_serializer, 'dummy')